+++ dst/app/routes/orders.js
@@ -594,12 +594,12 @@ ordersRouter.post('/findByConfirmationNumber', permitScopes_1.default(['orders.*
             // まだ注文が作成されていなければ、注文取引から検索するか検討中だが、いまのところ取引検索条件が足りない...
             throw new cinerino.factory.errors.NotFound('Order');
         }
-        if (USE_MULTI_ORDERS_BY_CONFIRMATION_NUMBER) {
-            res.json(orders);
-        }
-        else {
-            res.json(orders[0]);
-        }
+        res.json(orders);
+        // if (USE_MULTI_ORDERS_BY_CONFIRMATION_NUMBER) {
+        //     res.json(orders);
+        // } else {
+        //     res.json(orders[0]);
+        // }
     }
     catch (error) {
         next(error);