+++ dst/app/routes/payment/paymentCard.js
@@ -106,7 +106,6 @@ paymentCardPaymentRouter.post('/authorize', permitScopes_1.default(['transaction
     var _a;
     try {
         const actionRepo = new cinerino.repository.Action(mongoose.connection);
-        const projectRepo = new cinerino.repository.Project(mongoose.connection);
         const transactionRepo = new cinerino.repository.Transaction(mongoose.connection);
         let fromLocation = req.body.object.fromLocation;
         // トークン化された口座情報でリクエストされた場合、実口座情報へ変換する
@@ -202,23 +201,18 @@ paymentCardPaymentRouter.post('/authorize', permitScopes_1.default(['transaction
         //         };
         //     }
         // }
-        const currency = cinerino.factory.priceCurrency.JPY;
-        const action = yield cinerino.service.payment.paymentCard.authorize({
+        const action = yield cinerino.service.payment.chevre.authorize({
             project: req.project,
-            object: Object.assign(Object.assign({ 
-                // typeOf: cinerino.factory.paymentMethodType.PaymentCard,
-                typeOf: cinerino.factory.action.authorize.paymentMethod.any.ResultType.Payment, paymentMethod: (_a = req.body.object) === null || _a === void 0 ? void 0 : _a.paymentMethod, amount: Number(req.body.object.amount), currency: currency, additionalProperty: (Array.isArray(req.body.object.additionalProperty))
+            agent: { id: req.user.sub },
+            object: Object.assign(Object.assign({ typeOf: cinerino.factory.action.authorize.paymentMethod.any.ResultType.Payment, paymentMethod: (_a = req.body.object) === null || _a === void 0 ? void 0 : _a.paymentMethod, additionalProperty: (Array.isArray(req.body.object.additionalProperty))
                     ? req.body.object.additionalProperty.map((p) => {
                         return { name: String(p.name), value: String(p.value) };
                     })
-                    : [] }, (typeof req.body.object.name === 'string') ? { name: req.body.object.name } : undefined), (fromLocation !== undefined) ? { fromLocation } : {}
-            // ...(toLocation !== undefined) ? { toLocation } : {}
-            ),
-            agent: { id: req.user.sub },
-            purpose: { typeOf: req.body.purpose.typeOf, id: req.body.purpose.id }
+                    : [], amount: Number(req.body.object.amount), accountId: fromLocation.identifier }, (typeof req.body.object.name === 'string') ? { name: req.body.object.name } : undefined), (typeof req.body.object.description === 'string') ? { description: req.body.object.description } : undefined),
+            purpose: { typeOf: req.body.purpose.typeOf, id: req.body.purpose.id },
+            paymentServiceType: cinerino.factory.chevre.service.paymentService.PaymentServiceType.PaymentCard
         })({
             action: actionRepo,
-            project: projectRepo,
             transaction: transactionRepo
         });
         res.status(http_status_1.CREATED)
@@ -243,14 +237,13 @@ paymentCardPaymentRouter.put('/authorize/:actionId/void', permitScopes_1.default
     })(req, res, next);
 }), (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
     try {
-        yield cinerino.service.payment.paymentCard.voidTransaction({
-            project: req.project,
-            id: req.params.actionId,
+        yield cinerino.service.payment.chevre.voidPayment({
+            project: { id: req.project.id, typeOf: req.project.typeOf },
             agent: { id: req.user.sub },
+            id: req.params.actionId,
             purpose: { typeOf: req.body.purpose.typeOf, id: req.body.purpose.id }
         })({
             action: new cinerino.repository.Action(mongoose.connection),
-            project: new cinerino.repository.Project(mongoose.connection),
             transaction: new cinerino.repository.Transaction(mongoose.connection)
         });
         res.status(http_status_1.NO_CONTENT)