+++ src/app/routes/health.ts
@@ -18,9 +18,49 @@ const TIMEOUT_GIVE_UP_CHECKING_IN_MILLISECONDS = 3000;
 healthRouter.get(
     '',
     async (_, res, next) => {
+        let timer: NodeJS.Timer | undefined;
+
         try {
-            await mongoose.connection.db.admin()
-                .ping();
+            await new Promise(async (resolve, reject) => {
+                let givenUpChecking = false;
+
+                timer = setInterval(
+                    async () => {
+                        // すでにあきらめていたら何もしない
+                        if (givenUpChecking) {
+                            return;
+                        }
+
+                        if (typeof mongoose.connection?.db?.admin !== 'function') {
+                            return;
+                        }
+
+                        try {
+                            await mongoose.connection.db.admin()
+                                .ping();
+
+                            resolve();
+                        } catch (error) {
+                            reject(error);
+                        }
+                    },
+                    // tslint:disable-next-line:no-magic-numbers
+                    500
+                );
+
+                setTimeout(
+                    () => {
+                        givenUpChecking = true;
+                        reject(new Error('unable to check MongoDB connection'));
+                    },
+                    TIMEOUT_GIVE_UP_CHECKING_IN_MILLISECONDS
+                );
+            });
+
+            if (timer !== undefined) {
+                clearInterval(timer);
+            }
+
             await new Promise(async (resolve, reject) => {
                 let givenUpChecking = false;
 
@@ -52,10 +92,13 @@ healthRouter.get(
             res.status(OK)
                 .send('healthy!');
         } catch (error) {
+            if (timer !== undefined) {
+                clearInterval(timer);
+            }
+
             next(error);
         }
-    }
-);
+    });
 
 // healthRouter.get(
 //     '/closeMongo',