+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - ヘルスチェックを調整
+- 注文取引確定時の確認番号カスタム指定を削除
 
 ### Deprecated