+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - ヘルスチェックを調整
 - 注文取引確定時の確認番号カスタム指定を削除
 - update @chevre/api-nodejs-client
+- MovieTicket系統決済の場合、決済承認前に注文の確認番号を発行するように調整
 
 ### Deprecated