+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - update @chevre/api-nodejs-client
 - MovieTicket系統決済の場合、決済承認前に注文の確認番号を発行するように調整
 - アクション検索条件拡張
+- ペイメントカード決済処理をChevre決済取引に統合
 
 ### Deprecated