+++ CHANGELOG.md
@@ -18,6 +18,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v5.40.0 - 2020-11-13
+
+### Changed
+
+- ヘルスチェックを調整
+- 注文取引確定時の確認番号カスタム指定を削除
+- update @chevre/api-nodejs-client
+- MovieTicket系統決済の場合、決済承認前に注文の確認番号を発行するように調整
+- アクション検索条件拡張
+- ペイメントカード決済処理をChevre決済取引に統合
+- 確認番号での注文照会レスポンスをArrayに変更
+
 ## v5.39.3 - 2020-11-06
 
 ### Changed